Update pkg/configtx tests to use New to create ConfigTx (#1143)
And some miscellaneous cleanup elsewhere.

FAB-17775 #done
